requirements.txt

Django>=4.2
django-ckeditor
django-crispy-forms
crispy-bootstrap4
pytest
pytest-django